Split send_receive tab
