+++ applications/tari_console_wallet/src/ui/components/mod.rs
@@ -24,7 +24,8 @@ pub mod balance;
 pub mod base_node;
 mod component;
 pub mod network_tab;
-pub mod send_receive_tab;
+pub mod receive_tab;
+pub mod send_tab;
 pub mod tabs_container;
 pub mod transactions_tab;
