+++ applications/tari_console_wallet/src/ui/components/send_tab.rs
@@ -19,7 +19,7 @@ use tui::{
 };
 use unicode_width::UnicodeWidthStr;
 
-pub struct SendReceiveTab {
+pub struct SendTab {
     balance: Balance,
     send_input_mode: SendInputMode,
     edit_contact_mode: ContactInputMode,
@@ -38,7 +38,7 @@ pub struct SendReceiveTab {
     confirmation_dialog: Option<ConfirmationDialogType>,
 }
 
-impl SendReceiveTab {
+impl SendTab {
     pub fn new() -> Self {
         Self {
             balance: Balance::new(),
@@ -171,80 +171,6 @@ impl SendReceiveTab {
         }
     }
 
-    fn draw_whoami<B>(&self, f: &mut Frame<B>, area: Rect, app_state: &AppState)
-    where B: Backend {
-        let block = Block::default().borders(Borders::ALL).title(Span::styled(
-            "Who Am I?",
-            Style::default().fg(Color::White).add_modifier(Modifier::BOLD),
-        ));
-        f.render_widget(block, area);
-
-        let help_body_area = Layout::default()
-            .constraints([Constraint::Min(42)].as_ref())
-            .margin(1)
-            .split(area);
-
-        let chunks = Layout::default()
-            .direction(Direction::Horizontal)
-            .constraints([Constraint::Length(48), Constraint::Min(1)].as_ref())
-            .margin(1)
-            .split(help_body_area[0]);
-
-        let qr_code = Paragraph::new(app_state.get_identity().qr_code.as_str()).block(Block::default());
-        //.wrap(Wrap { trim: true });
-        f.render_widget(qr_code, chunks[0]);
-
-        let info_chunks = Layout::default()
-            .constraints(
-                [
-                    Constraint::Length(1), // Lining up fields with Qr Code
-                    Constraint::Length(3),
-                    Constraint::Length(3),
-                    Constraint::Length(3),
-                    Constraint::Min(1),
-                ]
-                .as_ref(),
-            )
-            .horizontal_margin(1)
-            .split(chunks[1]);
-
-        // Public Key
-        let block = Block::default()
-            .borders(Borders::ALL)
-            .title(Span::styled("Public Key", Style::default().fg(Color::White)));
-        f.render_widget(block, info_chunks[1]);
-        let label_layout = Layout::default()
-            .constraints([Constraint::Length(1)].as_ref())
-            .margin(1)
-            .split(info_chunks[1]);
-        let public_key = Paragraph::new(app_state.get_identity().public_key.as_str());
-        f.render_widget(public_key, label_layout[0]);
-
-        // Public Address
-        let block = Block::default()
-            .borders(Borders::ALL)
-            .title(Span::styled("Public Address", Style::default().fg(Color::White)));
-        f.render_widget(block, info_chunks[2]);
-        let label_layout = Layout::default()
-            .constraints([Constraint::Length(1)].as_ref())
-            .margin(1)
-            .split(info_chunks[2]);
-        let public_address = Paragraph::new(app_state.get_identity().public_address.as_str());
-        f.render_widget(public_address, label_layout[0]);
-
-        // Emoji ID
-        let block = Block::default()
-            .borders(Borders::ALL)
-            .title(Span::styled("Emoji ID", Style::default().fg(Color::White)));
-        f.render_widget(block, info_chunks[3]);
-        let label_layout = Layout::default()
-            .constraints([Constraint::Length(1)].as_ref())
-            .margin(1)
-            .split(info_chunks[3]);
-        let emoji_id = Paragraph::new(app_state.get_identity().emoji_id.as_str());
-        f.render_widget(emoji_id, label_layout[0]);
-    }
-
     fn draw_contacts<B>(&mut self, f: &mut Frame<B>, area: Rect, app_state: &AppState)
     where B: Backend {
         let block = Block::default().borders(Borders::ALL).title(Span::styled(
@@ -367,7 +293,7 @@ impl SendReceiveTab {
     }
 }
 
-impl<B: Backend> Component<B> for SendReceiveTab {
+impl<B: Backend> Component<B> for SendTab {
     fn draw(&mut self, f: &mut Frame<B>, area: Rect, app_state: &AppState) {
         let areas = Layout::default()
             .constraints(
@@ -389,9 +315,7 @@ impl<B: Backend> Component<B> for SendReceiveTab {
             if self.show_edit_contact {
                 self.draw_edit_contact(f, area, app_state);
             }
-        } else {
-            self.draw_whoami(f, areas[2], app_state);
-        }
+        };
 
         let rx_option = self.send_result_watch.take();
         if let Some(rx) = rx_option {
